src/requirements.txt
anyio==4.2.0
certifi==2024.8.30
exceptiongroup==1.2.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.8
psycopg2-binary==2.9.9
python-telegram-bot==20.8
sniffio==1.3.0
SQLAlchemy==2.0.32
typing_extensions==4.9.0
python-dotenv~=1.0.1